// src/libs/payment-gateway-inline/enums.ts
export enum PaymentGatewayInlineProvider {
  FLUTTERWAVE = "FLUTTERWAVE",
  PAYSTACK = "PAYSTACK",
  SEERBIT = "SEERBIT",
  REMITA = "REMITA",
}

export enum PaymentGatewayInlineChannel {
  CARD = "card",
  BANK = "bank",
  BANK_TRANSFER = "bank_transfer",
  USSD = "ussd",
  QR = "qr",
  MOBILE_MONEY = "mobile_money",
  WALLET = "wallet",
}